backend/requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.9
django-cors-headers>=3.10
 Pillow>=8.2
django-filter>=2.4